Reduce the number of functions needed to call to install functionality to scripts
