+++ Sprocket/Scripting/LuaLibrary.cpp
@@ -119,170 +119,6 @@ std::string view_source_code(std::string_view name)
     )lua", name);
 }
 
-}
-
-void load_entity_transformation_functions(lua::Script& script)
-{
-    lua_State* L = script.native_handle();
-
-    lua_register(L, "SetLookAt", [](lua_State* L) {
-        if (!check_arg_count(L, 3)) { return luaL_error(L, "Bad number of args"); }
-        auto& reg = *get_pointer<spkt::registry>(L, "__registry__");
-        auto entity = Converter<spkt::entity>::read(L, 1);
-        glm::vec3 p = Converter<glm::vec3>::read(L, 2);
-        glm::vec3 t = Converter<glm::vec3>::read(L, 3);
-        auto& tr = reg.get<Transform3DComponent>(entity);
-        tr.position = p;
-        tr.orientation = glm::conjugate(glm::quat_cast(glm::lookAt(tr.position, t, {0.0, 1.0, 0.0})));
-        return 0;
-    });
-
-    lua_register(L, "RotateY", [](lua_State* L) {
-        if (!check_arg_count(L, 2)) { return luaL_error(L, "Bad number of args"); };
-        auto& reg = *get_pointer<spkt::registry>(L, "__registry__");
-        auto entity = Converter<spkt::entity>::read(L, 1);
-        auto& tr = reg.get<Transform3DComponent>(entity);
-        float yaw = (float)lua_tonumber(L, 2);
-        tr.orientation = glm::rotate(tr.orientation, yaw, {0, 1, 0});
-        return 0;
-    });
-
-    lua_register(L, "GetForwardsDir", [](lua_State* L) {
-        if (!check_arg_count(L, 1)) { return luaL_error(L, "Bad number of args"); }
-        auto& reg = *get_pointer<spkt::registry>(L, "__registry__");
-        auto entity = Converter<spkt::entity>::read(L, 1);
-        auto& tr = reg.get<Transform3DComponent>(entity);
-        auto o = tr.orientation;
-
-        if (auto* pc = reg.get_if<Camera3DComponent>(entity)) {
-            o = glm::rotate(o, pc->pitch, {1, 0, 0});
-        }
-
-        Converter<glm::vec3>::push(L, Maths::Forwards(o));
-        return 1;
-    });
-
-    lua_register(L, "GetRightDir", [](lua_State* L) {
-        if (!check_arg_count(L, 1)) { return luaL_error(L, "Bad number of args"); }
-        auto& reg = *get_pointer<spkt::registry>(L, "__registry__");
-        auto entity = Converter<spkt::entity>::read(L, 1);
-        auto& tr = reg.get<Transform3DComponent>(entity);
-        Converter<glm::vec3>::push(L, Maths::Right(tr.orientation));
-        return 1;
-    });
-
-    lua_register(L, "MakeUpright", [](lua_State* L) {
-        if (!check_arg_count(L, 2)) { return luaL_error(L, "Bad number of args"); }
-        auto& reg = *get_pointer<spkt::registry>(L, "__registry__");
-        auto entity = Converter<spkt::entity>::read(L, 1);
-        auto& tr = reg.get<Transform3DComponent>(entity);
-        float yaw = Converter<float>::read(L, 2);
-        tr.orientation = glm::quat(glm::vec3(0, yaw, 0));
-        return 0;
-    });
-}
-
-void load_registry_functions(lua::Script& script, spkt::registry& registry)
-{
-    using namespace spkt;
-
-    lua_State* L = script.native_handle();
-    script.set_value("__registry__", &registry);
-
-    // Add functions for creating and destroying entities.
-    lua_register(L, "NewEntity", [](lua_State* L) {
-        if (!check_arg_count(L, 0)) { return luaL_error(L, "Bad number of args"); }
-        auto& reg = *get_pointer<spkt::registry>(L, "__registry__");
-        Converter<spkt::entity>::push(L, reg.create());
-        return 1;
-    });
-
-    lua_register(L, "DeleteEntity", [](lua_State* L) {
-        if (!check_arg_count(L, 1)) { return luaL_error(L, "Bad number of args"); }
-        auto& reg = *get_pointer<spkt::registry>(L, "__registry__");
-        auto entity = Converter<spkt::entity>::read(L, 1);
-        add_command(L, [&, entity]() { reg.destroy(entity); });
-        return 0;
-    });
-
-    // Add functions for iterating over all entities in __scene__. The C++ functions
-    // should not be used directly, instead they should be used via the Scene:Each
-    // function implemented last in Lua.
-    lua_register(L, "_view_All_init", view_init<>);
-    lua_register(L, "_view_All_next", view_next<>);
-    luaL_dostring(L, view_source_code("All").c_str());
-
-    // VIEW FUNCTIONS - GENERATED BY DATAMATIC
-    lua_register(L, "_view_NameComponent_init", view_init<NameComponent>);
-    lua_register(L, "_view_NameComponent_next", view_next<NameComponent>);
-    luaL_dostring(L, view_source_code("NameComponent").c_str());
-
-    lua_register(L, "_view_Transform2DComponent_init", view_init<Transform2DComponent>);
-    lua_register(L, "_view_Transform2DComponent_next", view_next<Transform2DComponent>);
-    luaL_dostring(L, view_source_code("Transform2DComponent").c_str());
-
-    lua_register(L, "_view_Transform3DComponent_init", view_init<Transform3DComponent>);
-    lua_register(L, "_view_Transform3DComponent_next", view_next<Transform3DComponent>);
-    luaL_dostring(L, view_source_code("Transform3DComponent").c_str());
-
-    lua_register(L, "_view_StaticModelComponent_init", view_init<StaticModelComponent>);
-    lua_register(L, "_view_StaticModelComponent_next", view_next<StaticModelComponent>);
-    luaL_dostring(L, view_source_code("StaticModelComponent").c_str());
-
-    lua_register(L, "_view_AnimatedModelComponent_init", view_init<AnimatedModelComponent>);
-    lua_register(L, "_view_AnimatedModelComponent_next", view_next<AnimatedModelComponent>);
-    luaL_dostring(L, view_source_code("AnimatedModelComponent").c_str());
-
-    lua_register(L, "_view_RigidBody3DComponent_init", view_init<RigidBody3DComponent>);
-    lua_register(L, "_view_RigidBody3DComponent_next", view_next<RigidBody3DComponent>);
-    luaL_dostring(L, view_source_code("RigidBody3DComponent").c_str());
-
-    lua_register(L, "_view_BoxCollider3DComponent_init", view_init<BoxCollider3DComponent>);
-    lua_register(L, "_view_BoxCollider3DComponent_next", view_next<BoxCollider3DComponent>);
-    luaL_dostring(L, view_source_code("BoxCollider3DComponent").c_str());
-
-    lua_register(L, "_view_SphereCollider3DComponent_init", view_init<SphereCollider3DComponent>);
-    lua_register(L, "_view_SphereCollider3DComponent_next", view_next<SphereCollider3DComponent>);
-    luaL_dostring(L, view_source_code("SphereCollider3DComponent").c_str());
-
-    lua_register(L, "_view_CapsuleCollider3DComponent_init", view_init<CapsuleCollider3DComponent>);
-    lua_register(L, "_view_CapsuleCollider3DComponent_next", view_next<CapsuleCollider3DComponent>);
-    luaL_dostring(L, view_source_code("CapsuleCollider3DComponent").c_str());
-
-    lua_register(L, "_view_ScriptComponent_init", view_init<ScriptComponent>);
-    lua_register(L, "_view_ScriptComponent_next", view_next<ScriptComponent>);
-    luaL_dostring(L, view_source_code("ScriptComponent").c_str());
-
-    lua_register(L, "_view_Camera3DComponent_init", view_init<Camera3DComponent>);
-    lua_register(L, "_view_Camera3DComponent_next", view_next<Camera3DComponent>);
-    luaL_dostring(L, view_source_code("Camera3DComponent").c_str());
-
-    lua_register(L, "_view_PathComponent_init", view_init<PathComponent>);
-    lua_register(L, "_view_PathComponent_next", view_next<PathComponent>);
-    luaL_dostring(L, view_source_code("PathComponent").c_str());
-
-    lua_register(L, "_view_LightComponent_init", view_init<LightComponent>);
-    lua_register(L, "_view_LightComponent_next", view_next<LightComponent>);
-    luaL_dostring(L, view_source_code("LightComponent").c_str());
-
-    lua_register(L, "_view_SunComponent_init", view_init<SunComponent>);
-    lua_register(L, "_view_SunComponent_next", view_next<SunComponent>);
-    luaL_dostring(L, view_source_code("SunComponent").c_str());
-
-    lua_register(L, "_view_AmbienceComponent_init", view_init<AmbienceComponent>);
-    lua_register(L, "_view_AmbienceComponent_next", view_next<AmbienceComponent>);
-    luaL_dostring(L, view_source_code("AmbienceComponent").c_str());
-
-    lua_register(L, "_view_ParticleComponent_init", view_init<ParticleComponent>);
-    lua_register(L, "_view_ParticleComponent_next", view_next<ParticleComponent>);
-    luaL_dostring(L, view_source_code("ParticleComponent").c_str());
-
-    lua_register(L, "_view_CollisionEvent_init", view_init<CollisionEvent>);
-    lua_register(L, "_view_CollisionEvent_next", view_next<CollisionEvent>);
-    luaL_dostring(L, view_source_code("CollisionEvent").c_str());
-
-}
-
 // COMPONENT RELATED CODE - GENERATED BY DATAMATIC
 // C++ Functions for NameComponent =====================================================
 
@@ -954,6 +790,67 @@ int _AddCollisionEvent(lua_State* L) {
 }
 
 
+void load_entity_transformation_functions(lua::Script& script)
+{
+    lua_State* L = script.native_handle();
+
+    lua_register(L, "SetLookAt", [](lua_State* L) {
+        if (!check_arg_count(L, 3)) { return luaL_error(L, "Bad number of args"); }
+        auto& reg = *get_pointer<spkt::registry>(L, "__registry__");
+        auto entity = Converter<spkt::entity>::read(L, 1);
+        glm::vec3 p = Converter<glm::vec3>::read(L, 2);
+        glm::vec3 t = Converter<glm::vec3>::read(L, 3);
+        auto& tr = reg.get<Transform3DComponent>(entity);
+        tr.position = p;
+        tr.orientation = glm::conjugate(glm::quat_cast(glm::lookAt(tr.position, t, {0.0, 1.0, 0.0})));
+        return 0;
+    });
+
+    lua_register(L, "RotateY", [](lua_State* L) {
+        if (!check_arg_count(L, 2)) { return luaL_error(L, "Bad number of args"); };
+        auto& reg = *get_pointer<spkt::registry>(L, "__registry__");
+        auto entity = Converter<spkt::entity>::read(L, 1);
+        auto& tr = reg.get<Transform3DComponent>(entity);
+        float yaw = (float)lua_tonumber(L, 2);
+        tr.orientation = glm::rotate(tr.orientation, yaw, {0, 1, 0});
+        return 0;
+    });
+
+    lua_register(L, "GetForwardsDir", [](lua_State* L) {
+        if (!check_arg_count(L, 1)) { return luaL_error(L, "Bad number of args"); }
+        auto& reg = *get_pointer<spkt::registry>(L, "__registry__");
+        auto entity = Converter<spkt::entity>::read(L, 1);
+        auto& tr = reg.get<Transform3DComponent>(entity);
+        auto o = tr.orientation;
+
+        if (auto* pc = reg.get_if<Camera3DComponent>(entity)) {
+            o = glm::rotate(o, pc->pitch, {1, 0, 0});
+        }
+
+        Converter<glm::vec3>::push(L, Maths::Forwards(o));
+        return 1;
+    });
+
+    lua_register(L, "GetRightDir", [](lua_State* L) {
+        if (!check_arg_count(L, 1)) { return luaL_error(L, "Bad number of args"); }
+        auto& reg = *get_pointer<spkt::registry>(L, "__registry__");
+        auto entity = Converter<spkt::entity>::read(L, 1);
+        auto& tr = reg.get<Transform3DComponent>(entity);
+        Converter<glm::vec3>::push(L, Maths::Right(tr.orientation));
+        return 1;
+    });
+
+    lua_register(L, "MakeUpright", [](lua_State* L) {
+        if (!check_arg_count(L, 2)) { return luaL_error(L, "Bad number of args"); }
+        auto& reg = *get_pointer<spkt::registry>(L, "__registry__");
+        auto entity = Converter<spkt::entity>::read(L, 1);
+        auto& tr = reg.get<Transform3DComponent>(entity);
+        float yaw = Converter<float>::read(L, 2);
+        tr.orientation = glm::quat(glm::vec3(0, yaw, 0));
+        return 0;
+    });
+}
+
 void load_entity_component_functions(lua::Script& script)
 {
     lua_State* L = script.native_handle();
@@ -1622,6 +1519,130 @@ void load_entity_component_functions(lua::Script& script)
     lua_register(L, "HasCollisionEvent", &has_impl<CollisionEvent>);
     lua_register(L, "RemoveCollisionEvent", &remove_impl<CollisionEvent>);
 
+
+}
+
+void load_entity_functions(lua::Script& script)
+{
+    lua_State* L = script.native_handle();
+
+    luaL_newmetatable(L, "entity_id");
+
+    lua_pushcfunction(L, [](lua_State* L) {
+        spkt::entity* self = (spkt::entity*)luaL_checkudata(L, 1, "entity_id");
+        spkt::entity* other = (spkt::entity*)luaL_checkudata(L, 2, "entity_id");
+        Converter<bool>::push(L, *self == *other);
+        return 1;
+    });
+    lua_setfield(L, -2, "__eq");
+
+    lua_setglobal(L, "entity_id");
+}
+
+}
+
+void load_registry(lua::Script& script, spkt::registry& registry)
+{
+    using namespace spkt;
+
+    lua_State* L = script.native_handle();
+    script.set_value("__registry__", &registry);
+
+    load_entity_transformation_functions(script);
+    load_entity_component_functions(script);
+    load_entity_functions(script);
+
+    // Add functions for creating and destroying entities.
+    lua_register(L, "NewEntity", [](lua_State* L) {
+        if (!check_arg_count(L, 0)) { return luaL_error(L, "Bad number of args"); }
+        auto& reg = *get_pointer<spkt::registry>(L, "__registry__");
+        Converter<spkt::entity>::push(L, reg.create());
+        return 1;
+    });
+
+    lua_register(L, "DeleteEntity", [](lua_State* L) {
+        if (!check_arg_count(L, 1)) { return luaL_error(L, "Bad number of args"); }
+        auto& reg = *get_pointer<spkt::registry>(L, "__registry__");
+        auto entity = Converter<spkt::entity>::read(L, 1);
+        add_command(L, [&, entity]() { reg.destroy(entity); });
+        return 0;
+    });
+
+    // Add functions for iterating over all entities in __scene__. The C++ functions
+    // should not be used directly, instead they should be used via the Scene:Each
+    // function implemented last in Lua.
+    lua_register(L, "_view_All_init", view_init<>);
+    lua_register(L, "_view_All_next", view_next<>);
+    luaL_dostring(L, view_source_code("All").c_str());
+
+    // VIEW FUNCTIONS - GENERATED BY DATAMATIC
+    lua_register(L, "_view_NameComponent_init", view_init<NameComponent>);
+    lua_register(L, "_view_NameComponent_next", view_next<NameComponent>);
+    luaL_dostring(L, view_source_code("NameComponent").c_str());
+
+    lua_register(L, "_view_Transform2DComponent_init", view_init<Transform2DComponent>);
+    lua_register(L, "_view_Transform2DComponent_next", view_next<Transform2DComponent>);
+    luaL_dostring(L, view_source_code("Transform2DComponent").c_str());
+
+    lua_register(L, "_view_Transform3DComponent_init", view_init<Transform3DComponent>);
+    lua_register(L, "_view_Transform3DComponent_next", view_next<Transform3DComponent>);
+    luaL_dostring(L, view_source_code("Transform3DComponent").c_str());
+
+    lua_register(L, "_view_StaticModelComponent_init", view_init<StaticModelComponent>);
+    lua_register(L, "_view_StaticModelComponent_next", view_next<StaticModelComponent>);
+    luaL_dostring(L, view_source_code("StaticModelComponent").c_str());
+
+    lua_register(L, "_view_AnimatedModelComponent_init", view_init<AnimatedModelComponent>);
+    lua_register(L, "_view_AnimatedModelComponent_next", view_next<AnimatedModelComponent>);
+    luaL_dostring(L, view_source_code("AnimatedModelComponent").c_str());
+
+    lua_register(L, "_view_RigidBody3DComponent_init", view_init<RigidBody3DComponent>);
+    lua_register(L, "_view_RigidBody3DComponent_next", view_next<RigidBody3DComponent>);
+    luaL_dostring(L, view_source_code("RigidBody3DComponent").c_str());
+
+    lua_register(L, "_view_BoxCollider3DComponent_init", view_init<BoxCollider3DComponent>);
+    lua_register(L, "_view_BoxCollider3DComponent_next", view_next<BoxCollider3DComponent>);
+    luaL_dostring(L, view_source_code("BoxCollider3DComponent").c_str());
+
+    lua_register(L, "_view_SphereCollider3DComponent_init", view_init<SphereCollider3DComponent>);
+    lua_register(L, "_view_SphereCollider3DComponent_next", view_next<SphereCollider3DComponent>);
+    luaL_dostring(L, view_source_code("SphereCollider3DComponent").c_str());
+
+    lua_register(L, "_view_CapsuleCollider3DComponent_init", view_init<CapsuleCollider3DComponent>);
+    lua_register(L, "_view_CapsuleCollider3DComponent_next", view_next<CapsuleCollider3DComponent>);
+    luaL_dostring(L, view_source_code("CapsuleCollider3DComponent").c_str());
+
+    lua_register(L, "_view_ScriptComponent_init", view_init<ScriptComponent>);
+    lua_register(L, "_view_ScriptComponent_next", view_next<ScriptComponent>);
+    luaL_dostring(L, view_source_code("ScriptComponent").c_str());
+
+    lua_register(L, "_view_Camera3DComponent_init", view_init<Camera3DComponent>);
+    lua_register(L, "_view_Camera3DComponent_next", view_next<Camera3DComponent>);
+    luaL_dostring(L, view_source_code("Camera3DComponent").c_str());
+
+    lua_register(L, "_view_PathComponent_init", view_init<PathComponent>);
+    lua_register(L, "_view_PathComponent_next", view_next<PathComponent>);
+    luaL_dostring(L, view_source_code("PathComponent").c_str());
+
+    lua_register(L, "_view_LightComponent_init", view_init<LightComponent>);
+    lua_register(L, "_view_LightComponent_next", view_next<LightComponent>);
+    luaL_dostring(L, view_source_code("LightComponent").c_str());
+
+    lua_register(L, "_view_SunComponent_init", view_init<SunComponent>);
+    lua_register(L, "_view_SunComponent_next", view_next<SunComponent>);
+    luaL_dostring(L, view_source_code("SunComponent").c_str());
+
+    lua_register(L, "_view_AmbienceComponent_init", view_init<AmbienceComponent>);
+    lua_register(L, "_view_AmbienceComponent_next", view_next<AmbienceComponent>);
+    luaL_dostring(L, view_source_code("AmbienceComponent").c_str());
+
+    lua_register(L, "_view_ParticleComponent_init", view_init<ParticleComponent>);
+    lua_register(L, "_view_ParticleComponent_next", view_next<ParticleComponent>);
+    luaL_dostring(L, view_source_code("ParticleComponent").c_str());
+
+    lua_register(L, "_view_CollisionEvent_init", view_init<CollisionEvent>);
+    lua_register(L, "_view_CollisionEvent_next", view_next<CollisionEvent>);
+    luaL_dostring(L, view_source_code("CollisionEvent").c_str());
 
 }
