+++ Sprocket/Scripting/LuaLibrary.dm.cpp
@@ -119,7 +119,42 @@ std::string view_source_code(std::string_view name)
     )lua", name);
 }
 
+// COMPONENT RELATED CODE - GENERATED BY DATAMATIC
+DATAMATIC_BEGIN SCRIPTABLE=true
+// C++ Functions for {{Comp::name}} =====================================================
+
+int _Get{{Comp::name}}(lua_State* L) {
+    if (!check_arg_count(L, 1)) { return luaL_error(L, "Bad number of args"); }
+    auto& reg = *get_pointer<spkt::registry>(L, "__registry__");
+    auto e = Converter<spkt::entity>::read(L, 1);
+    const auto& c = reg.get<{{Comp::name}}>(e);
+    Converter<{{Attr::type}}>::push(L, c.{{Attr::name}});
+    return {{Comp::attr_count}};
 }
+
+int _Set{{Comp::name}}(lua_State* L) {
+    if (!check_arg_count(L, {{Comp::attr_count}} + 1)) { return luaL_error(L, "Bad number of args"); }
+    int ptr = 0;
+    auto& reg = *get_pointer<spkt::registry>(L, "__registry__");
+    auto e = Converter<spkt::entity>::read(L, ++ptr);
+    auto& c = reg.get<{{Comp::name}}>(e);
+    c.{{Attr::name}} = Converter<{{Attr::type}}>::read(L, ++ptr);
+    return 0;
+}
+
+int _Add{{Comp::name}}(lua_State* L) {
+    if (!check_arg_count(L, {{Comp::attr_count}} + 1)) { return luaL_error(L, "Bad number of args"); }
+    int ptr = 0;
+    auto& reg = *get_pointer<spkt::registry>(L, "__registry__");
+    auto e = Converter<spkt::entity>::read(L, ++ptr);
+    assert(!reg.has<{{Comp::name}}>(e));
+    {{Comp::name}} c;
+    c.{{Attr::name}} = Converter<{{Attr::type}}>::read(L, ++ptr);
+    add_command(L, [&, e, c]() { reg.add<{{Comp::name}}>(e, c); });
+    return 0;
+}
+
+DATAMATIC_END
 
 void load_entity_transformation_functions(lua::Script& script)
 {
@@ -182,82 +217,6 @@ void load_entity_transformation_functions(lua::Script& script)
     });
 }
 
-void load_registry_functions(lua::Script& script, spkt::registry& registry)
-{
-    using namespace spkt;
-
-    lua_State* L = script.native_handle();
-    script.set_value("__registry__", &registry);
-
-    // Add functions for creating and destroying entities.
-    lua_register(L, "NewEntity", [](lua_State* L) {
-        if (!check_arg_count(L, 0)) { return luaL_error(L, "Bad number of args"); }
-        auto& reg = *get_pointer<spkt::registry>(L, "__registry__");
-        Converter<spkt::entity>::push(L, reg.create());
-        return 1;
-    });
-
-    lua_register(L, "DeleteEntity", [](lua_State* L) {
-        if (!check_arg_count(L, 1)) { return luaL_error(L, "Bad number of args"); }
-        auto& reg = *get_pointer<spkt::registry>(L, "__registry__");
-        auto entity = Converter<spkt::entity>::read(L, 1);
-        add_command(L, [&, entity]() { reg.destroy(entity); });
-        return 0;
-    });
-
-    // Add functions for iterating over all entities in __scene__. The C++ functions
-    // should not be used directly, instead they should be used via the Scene:Each
-    // function implemented last in Lua.    
-    lua_register(L, "_view_All_init", view_init<>);
-    lua_register(L, "_view_All_next", view_next<>);
-    luaL_dostring(L, view_source_code("All").c_str());
-
-    // VIEW FUNCTIONS - GENERATED BY DATAMATIC
-DATAMATIC_BEGIN SCRIPTABLE=true
-    lua_register(L, "_view_{{Comp::name}}_init", view_init<{{Comp::name}}>);
-    lua_register(L, "_view_{{Comp::name}}_next", view_next<{{Comp::name}}>);
-    luaL_dostring(L, view_source_code("{{Comp::name}}").c_str());
-
-DATAMATIC_END
-}
-
-// COMPONENT RELATED CODE - GENERATED BY DATAMATIC
-DATAMATIC_BEGIN SCRIPTABLE=true
-// C++ Functions for {{Comp::name}} =====================================================
-
-int _Get{{Comp::name}}(lua_State* L) {
-    if (!check_arg_count(L, 1)) { return luaL_error(L, "Bad number of args"); }
-    auto& reg = *get_pointer<spkt::registry>(L, "__registry__");
-    auto e = Converter<spkt::entity>::read(L, 1);
-    const auto& c = reg.get<{{Comp::name}}>(e);
-    Converter<{{Attr::type}}>::push(L, c.{{Attr::name}});
-    return {{Comp::attr_count}};
-}
-
-int _Set{{Comp::name}}(lua_State* L) {
-    if (!check_arg_count(L, {{Comp::attr_count}} + 1)) { return luaL_error(L, "Bad number of args"); }
-    int ptr = 0;
-    auto& reg = *get_pointer<spkt::registry>(L, "__registry__");
-    auto e = Converter<spkt::entity>::read(L, ++ptr);
-    auto& c = reg.get<{{Comp::name}}>(e);
-    c.{{Attr::name}} = Converter<{{Attr::type}}>::read(L, ++ptr);
-    return 0;
-}
-
-int _Add{{Comp::name}}(lua_State* L) {
-    if (!check_arg_count(L, {{Comp::attr_count}} + 1)) { return luaL_error(L, "Bad number of args"); }
-    int ptr = 0;
-    auto& reg = *get_pointer<spkt::registry>(L, "__registry__");
-    auto e = Converter<spkt::entity>::read(L, ++ptr);
-    assert(!reg.has<{{Comp::name}}>(e));
-    {{Comp::name}} c;
-    c.{{Attr::name}} = Converter<{{Attr::type}}>::read(L, ++ptr);
-    add_command(L, [&, e, c]() { reg.add<{{Comp::name}}>(e, c); });
-    return 0;
-}
-
-DATAMATIC_END
-
 void load_entity_component_functions(lua::Script& script)
 {
     lua_State* L = script.native_handle();
@@ -299,6 +258,68 @@ DATAMATIC_BEGIN SCRIPTABLE=true
     lua_register(L, "Has{{Comp::name}}", &has_impl<{{Comp::name}}>);
     lua_register(L, "Remove{{Comp::name}}", &remove_impl<{{Comp::name}}>);
 
+
+DATAMATIC_END
+}
+
+void load_entity_functions(lua::Script& script)
+{
+    lua_State* L = script.native_handle();
+
+    luaL_newmetatable(L, "entity_id");
+
+    lua_pushcfunction(L, [](lua_State* L) {
+        spkt::entity* self = (spkt::entity*)luaL_checkudata(L, 1, "entity_id");
+        spkt::entity* other = (spkt::entity*)luaL_checkudata(L, 2, "entity_id");
+        Converter<bool>::push(L, *self == *other);
+        return 1;
+    });
+    lua_setfield(L, -2, "__eq");
+
+    lua_setglobal(L, "entity_id");
+}
+
+}
+
+void load_registry(lua::Script& script, spkt::registry& registry)
+{
+    using namespace spkt;
+
+    lua_State* L = script.native_handle();
+    script.set_value("__registry__", &registry);
+
+    load_entity_transformation_functions(script);
+    load_entity_component_functions(script);
+    load_entity_functions(script);
+
+    // Add functions for creating and destroying entities.
+    lua_register(L, "NewEntity", [](lua_State* L) {
+        if (!check_arg_count(L, 0)) { return luaL_error(L, "Bad number of args"); }
+        auto& reg = *get_pointer<spkt::registry>(L, "__registry__");
+        Converter<spkt::entity>::push(L, reg.create());
+        return 1;
+    });
+
+    lua_register(L, "DeleteEntity", [](lua_State* L) {
+        if (!check_arg_count(L, 1)) { return luaL_error(L, "Bad number of args"); }
+        auto& reg = *get_pointer<spkt::registry>(L, "__registry__");
+        auto entity = Converter<spkt::entity>::read(L, 1);
+        add_command(L, [&, entity]() { reg.destroy(entity); });
+        return 0;
+    });
+
+    // Add functions for iterating over all entities in __scene__. The C++ functions
+    // should not be used directly, instead they should be used via the Scene:Each
+    // function implemented last in Lua.    
+    lua_register(L, "_view_All_init", view_init<>);
+    lua_register(L, "_view_All_next", view_next<>);
+    luaL_dostring(L, view_source_code("All").c_str());
+
+    // VIEW FUNCTIONS - GENERATED BY DATAMATIC
+DATAMATIC_BEGIN SCRIPTABLE=true
+    lua_register(L, "_view_{{Comp::name}}_init", view_init<{{Comp::name}}>);
+    lua_register(L, "_view_{{Comp::name}}_next", view_next<{{Comp::name}}>);
+    luaL_dostring(L, view_source_code("{{Comp::name}}").c_str());
 
 DATAMATIC_END
 }
