+++ Sprocket/Scripting/LuaLibrary.h
@@ -1,9 +1,5 @@
 #pragma once
 #include <Sprocket/Scene/ecs.h>
-// Functions for loading C++ implemented Lua functions into the given
-// lua state.
-
-struct lua_State;
 
 namespace spkt {
 namespace lua {
@@ -14,14 +10,7 @@ class Script;
 // creating/deleting entities as well as iterating entities. Also provides functions
 // for access the singleton entity, which in turn gives access to the keyboard, mouse
 // and window via the InputSingleton.
-void load_registry_functions(lua::Script& script, spkt::registry& registry);
-
-// Loads a bunch of helper functions to ease entity manipulation.
-void load_entity_transformation_functions(lua::Script& script);
-
-// Loads functions for accessing and modifying components on entities. The implementation
-// of this is generated by Datamatic.
-void load_entity_component_functions(lua::Script& script);
+void load_registry(lua::Script& script, spkt::registry& registry);
 
 }
 }
