+++ Sprocket/Scripting/LuaMaths.h
@@ -1,15 +1,12 @@
 #pragma once
-
-struct lua_State;
+#include <Sprocket/Scripting/LuaScript.h>
 
 namespace spkt {
 namespace lua {
 
 class Script;
 
-void load_vec3_functions(lua::Script& script);
-void load_vec2_functions(lua::Script& script);
-void load_entity_functions(lua::Script& script);
+void load_maths(lua::Script& script);
 
 }
 }
